requirements.txt
asgiref==3.8.1
Django==5.2.1
django-allauth==65.9.0
django-crispy-forms==2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
PyJWT==2.9.0
sqlparse==0.5.3
